+++ src/main/java/az/hrportal/hrportalapi/dto/position/request/SalaryRequestDto.java
@@ -6,12 +6,10 @@ import lombok.Setter;
 import lombok.ToString;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @ToString
 public class SalaryRequestDto {
-    BigDecimal salary;
+    Float salary;
 }
